changed dto and model

+++ src/main/kotlin/com/makarytskyi/rentcar/dto/car/UpdateCarRequest.kt
@@ -2,10 +2,11 @@ package com.makarytskyi.rentcar.dto.car
 
 import com.makarytskyi.rentcar.model.MongoCar
 import jakarta.validation.constraints.Min
+import java.math.BigDecimal
 
 data class UpdateCarRequest(
     @field:Min(0)
-    val price: Int?,
+    val price: BigDecimal?,
     val color: MongoCar.CarColor?,
 ) {
