+++ src/main/kotlin/com/makarytskyi/rentcar/dto/repairing/RepairingResponse.kt
@@ -1,13 +1,14 @@
 package com.makarytskyi.rentcar.dto.repairing
 
 import com.makarytskyi.rentcar.model.MongoRepairing
+import java.math.BigDecimal
 import java.util.Date
 
 data class RepairingResponse(
     val id: String,
     val carId: String,
     val date: Date?,
-    val price: Int?,
+    val price: BigDecimal?,
     val status: MongoRepairing.RepairingStatus?,
 ) {
