+++ src/main/kotlin/com/makarytskyi/rentcar/dto/repairing/UpdateRepairingRequest.kt
@@ -3,10 +3,11 @@ package com.makarytskyi.rentcar.dto.repairing
 import com.makarytskyi.rentcar.model.MongoRepairing
 import com.makarytskyi.rentcar.model.MongoRepairing.RepairingStatus
 import jakarta.validation.constraints.Min
+import java.math.BigDecimal
 
 data class UpdateRepairingRequest(
     @field:Min(0)
-    val price: Int?,
+    val price: BigDecimal?,
     val status: RepairingStatus?,
 ) {
